microservices/review_service/src/config/config.py
"""
Configuration module for Review Service
"""
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    """Application configuration"""
    
    # Flask Configuration
    DEBUG = os.getenv('FLASK_ENV', 'development') == 'development'
    PORT = int(os.getenv('PORT', 5002))
    SERVICE_NAME = os.getenv('SERVICE_NAME', 'review-service')
    
    # RabbitMQ Configuration
    RABBITMQ_HOST = os.getenv('RABBITMQ_HOST', 'localhost')
    RABBITMQ_PORT = int(os.getenv('RABBITMQ_PORT', 5672))
    RABBITMQ_USER = os.getenv('RABBITMQ_USER', 'guest')
    RABBITMQ_PASS = os.getenv('RABBITMQ_PASS', 'guest')
    RABBITMQ_VHOST = os.getenv('RABBITMQ_VHOST', '/')
    
    # RabbitMQ Queues and Exchanges
    REVIEW_QUEUE = 'review_notifications'
    REVIEW_EXCHANGE = 'review_exchange'
    
    # MongoDB Configuration
    MONGO_URI = os.getenv('MONGO_URI', 'mongodb://localhost:27017/')
    MONGO_DB = os.getenv('MONGO_DB', 'review_db')
    
    # User Service URL (for REST communication)
    USER_SERVICE_URL = os.getenv('USER_SERVICE_URL', 'http://localhost:8080')
    
    # Security
    JWT_SECRET = os.getenv('JWT_SECRET', 'your-secret-key-change-in-production')
    JWT_ALGORITHM = 'HS256'



